pyproject.toml
[tool.poetry]
name = "ai-crypto-hedge-fund-manager"
version = "0.1.0"
description = "A multi-agent system for cryptocurrency trading signals and automated swaps"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "crypto_project", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
requests = "^2.28.2"
python-dotenv = "^1.0.0"
pandas = "^2.0.0"
matplotlib = "^3.7.0"
pytest = "^7.3.1"
numpy = "^2.2.4"
uagents = "^0.22.1"
backoff = "^2.2.1"
newsapi-python = "^0.2.7"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 